requirements.txt
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
Django==3.1
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-modeltranslation==0.15.1
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
idna==2.10
model-bakery==1.1.0
outline-api==0.0.1
psycopg2==2.8.6
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
unicodecsv==0.14.1
urllib3==1.26.2
